Feats Venta
Logica de venta, con egreso de stock
Decoradores de sesion
Decoradores de transaccion

+++ src/main/java/com/tpi_pais/mega_store/products/dto/DetalleVentaDTO.java
@@ -2,6 +2,8 @@ package com.tpi_pais.mega_store.products.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DetalleVentaDTO {
     private Integer id;
@@ -9,6 +11,6 @@ public class DetalleVentaDTO {
     private Integer idProducto;
     private String producto;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
 }
